move ownKeys
